Refactoring of the TigerRAG package to follow best practices
