+++ TigerRag/demos/movie_recs/demo_ebr.py
@@ -1,8 +1,8 @@
-from tigerrag.rag.base.loaders import TigerRAGDataFrameLoader
-from tigerrag.rag.base.models import EmbeddingModel, TigerRAGEmbeddingModel
-from tigerrag.rag.ebr.similarity_search import FaissFlatL2Search
-
 from demo_utils import calculate_averaged_recall, calculate_single_recall
+
+from tigerrag.base.loaders import TigerRAGDataFrameLoader
+from tigerrag.base.models import EmbeddingModel, TigerRAGEmbeddingModel
+from tigerrag.ebr.similarity_search import FaissFlatL2Search
 
 """
 The provided Python script leverages the TigerRAG library and FAISS for an embeddings-based retrieval (EBR) system.
