+++ TigerRag/setup.py
@@ -1,4 +1,5 @@
 from pathlib import Path
+
 from setuptools import find_packages, setup
 
 with open(Path(__file__).absolute().parents[0] / "tigerrag" / "VERSION") as _f:
@@ -25,11 +26,16 @@ setup(
         "Programming Language :: Python :: 3.10",
         "Operating System :: OS Independent",
     ],
-    include_package_data=True,
-    package_data={
-        "tigerrag": ["demo/movie_recs/*.csv"]
-    },
-    install_requires=["bson >= 0.5.10"],
+    install_requires=[
+        "bson >= 0.5.10",
+        "pandas>=2.0.0",
+        "numpy>=1.22.0",
+        "torch>=2.1.0",
+        "transformers>=4.35.0",
+        "faiss-cpu>=1.7.0",
+        "scikit-learn>=1.2.0",
+        "openai>=0.28.0",
+    ],
     extras_require={
         "dev": ["pytest>=6.0", "twine>=3.0"],
     },
